src/requirements.txt
pytest>=3.0.7
logzero>=1.7.0
mysql-connector-python>=8.0.23
paramiko>=2.7.2